Gantt Chart for Task 1
We have given ourselves at least 2 days to complete each task. Between
the first set of tasks and our second, we have 4 days to email Dave,
review each others tasks and finalise our tasks. This goes for task 2 as
well. We will have the whole duration to also complete the Progress
Report, which we will be updating as we complete out tasks.